`dirext`: Add `new_linkable_file()` API
This helps support the key use case of "atomically replace an
existing file with new contents", as occurs in many places.  For
example in bootupd:
https://github.com/coreos/bootupd/blob/999c95121074dab9e0b889d044c417d1ed901148/src/backend/statefile.rs#L101

Basically our "database" is just a JSON file.  We don't want to use
e.g. `std::fs::write` because that could leave a corrupted half-written
file if the system crashes in the middle.