Dir: Add `open_dir_of` API
When porting to cap-std for many codebases, I think it will be
a fairly common operation to take an ambient path (which could be
absolute or relative to cwd) and want to access its parent directory.

In particular, in order to use this crate's handy "atomic file replacement"
APIs, one needs to correctly split such a path into a `Dir` instance
and a filename.