Drop `LinkableTempFile`, use `cap_tempfile::TempFile`
Now that all the core logic has been drained into
`cap_tempfile::TempFile` in https://github.com/bytecodealliance/cap-std/pull/239
we can just use it directly.

NOTE: This changes the semantics of some of the APIs; the previous
`replace()` API tried to do "reuse previous file mode if it existed"
but I think that's just a confusing trap.  Anything writing the file
in the general case needs to know the permissions it wants to use.