Add a fs_utf8 feature
I'd like to potentially start making use of the `fs_utf8::Dir`
in a lot of code in e.g. bootc/ostree related things because
ostree in particular hard requires UTF-8 filenames, and
we have redundant verifications.

This makes use of the new `as_cap_std()` API that landed
in cap-std.

Signed-off-by: Colin Walters <walters@verbum.org>