cmdext: Fix bug when srcfd == targetfd
See https://github.com/ostreedev/ostree-rs-ext/issues/664

Basically in the case when the source fd number is the same
as the target, `dup2` is a no-op. This is an astoundingly evil
bug because it means we just don't pass the expected
fd to the child process.

Fix this by detecting this situation and just stripping off
`O_CLOEXEC`.

Signed-off-by: Colin Walters <walters@verbum.org>