Add a `walk` method
There's a ton of use cases for a recursive directory walk, and
we end up re-implementing it over and over.

Unlike the `walkdir` crate, this runs via "internal
iteration" that accepts a closure which returns `std::ops::ControlFlow`
to ensure the caller can break out of iteration.

There's a config option that supports:

- Not traversing mount points
- Sorting entries