dirext: improve permissions handling in atomic_*
`atomic_replace_with` now takes the target permissions as an Option.

Existing permissions copy now happen after flushing the buffers,
so security related bits are not removed by the kernel.

When using filesystems that do not support `O_TEMPFILE`, we now change
the tempfile permissions to 0o000 as soon as possible.
A malicious process still has a really short time to `open()` the tempfile,
to prevent that for now you must the a restrictive umask.

Signed-off-by: Etienne Champetier <e.champetier@ateme.com>